Separate out the "PinCodeModal" component
